refactor: update the overall structure of the blog posts management interface
