+++ src/app/admin/page.module.scss
@@ -9,6 +9,20 @@
   width: 100vw;
 }
 
+.Section {
+  display: flex;
+  margin: 2rem;
+  text-transform: uppercase;
+}
+
+.SectionLinkItem {
+  display: flex;
+  border: 2px solid black;
+  padding: 3rem;
+  border-radius: 5px;
+  cursor: pointer;
+}
+
 .BlogPostCreate {
   display: flex;
   flex-direction: row;
