+++ src/app/admin/post/create/page.module.scss
@@ -1,0 +1,10 @@
+@import "../../../variables.module.scss";
+
+.Main {
+  display: flex;
+  flex-direction: column;
+  background-color: #eeeeee;
+  color: $background-color;
+  min-height: 100vh;
+  width: 100vw;
+}
