+++ src/app/admin/post/create/page.tsx
@@ -2,7 +2,6 @@ import styles from "./page.module.scss";
 import Link from "next/link";
 import Navbar from "../../../components/Navbar";
 import Writer from "@/app/components/Writer";
-import { MY_SITE_URL } from "../../../utils/constants";
 
 export type PostData = {
   postStatus: String;
@@ -11,19 +10,16 @@ export type PostData = {
   postCategory: String;
 };
 
-const PostUpdatePage = async ({ params }: { params: { slug: string } }) => {
-  const data = await fetch(`${MY_SITE_URL}/api/post?slug=${params.slug}`);
-  const postData = await data.json();
-
+const PostPage = async () => {
   return (
-    <section className={styles.Main}>
+    <div className={styles.Main}>
       <Navbar />
       <Link className={styles.Link} href={"/admin/post"}>
         Back
       </Link>
-      <Writer postData={postData.data} />
-    </section>
+      <Writer />
+    </div>
   );
 };
 
-export default PostUpdatePage;
+export default PostPage;
